requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astroid==3.0.0
cffi==1.16.0
colorama==0.4.6
cryptography==41.0.4
dill==0.3.7
isort==5.12.0
mccabe==0.7.0
platformdirs==3.11.0
pycparser==2.21
pylint==3.0.1
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
